Extract actions from button components
